Spinlock
Changed from long to int.

+++ code/foundation/threading/spinlock.h
@@ -69,7 +69,7 @@ Spinlock::operator=(Spinlock&& rhs)
         return;
 
     // Lock this thread until whoever is holding it returns the lock
-    while (Interlocked::CompareExchange((volatile long*)&this->lock, InvalidThreadId, InvalidThreadId) != InvalidThreadId)
+    while (Interlocked::CompareExchange((volatile int*)&this->lock, InvalidThreadId, InvalidThreadId) != InvalidThreadId)
     {
         Thread::YieldThread();
     }
@@ -89,7 +89,7 @@ Spinlock::Lock()
         return;
 
     // Otherwise, enter the spin to exchange the thread id to ours
-    while (Interlocked::CompareExchange((volatile long*) &this->lock, threadId, InvalidThreadId) != InvalidThreadId)
+    while (Interlocked::CompareExchange((volatile int*) &this->lock, threadId, InvalidThreadId) != InvalidThreadId)
     {
         Thread::YieldThread();
     }
@@ -102,7 +102,7 @@ inline void
 Spinlock::Unlock()
 {
     n_assert(this->lock != InvalidThreadId);
-    Threading::Interlocked::Exchange((volatile long*) &this->lock, InvalidThreadId);
+    Threading::Interlocked::Exchange((volatile int*) &this->lock, InvalidThreadId);
 }
 
 struct SpinlockScope
